test(runtime-policy): stabilize expected-error stdin (#644)
